add common module and move models to it
