+++ rohmu/object_storage/azure.py
@@ -5,7 +5,7 @@ Copyright (c) 2016 Ohmu Ltd
 Copyright (c) 2022 Aiven, Helsinki, Finland. https://aiven.io/
 See LICENSE for details
 """
-from ..models import ProxyInfo, StorageModel
+from ..common.models import ProxyInfo, StorageModel
 from ..notifier.interface import Notifier
 from .base import IncrementalProgressCallbackType, ProgressProportionCallbackType
