+++ rohmu/object_storage/local.py
@@ -5,8 +5,8 @@ Copyright (c) 2016 Ohmu Ltd
 Copyright (c) 2022 Aiven, Helsinki, Finland. https://aiven.io/
 See LICENSE for details
 """
+from ..common.models import StorageModel
 from ..errors import FileNotFoundFromStorageError, LocalFileIsRemoteFileError
-from ..models import StorageModel
 from ..notifier.interface import Notifier
 from .base import (
     BaseTransfer,
