+++ rohmu/object_storage/sftp.py
@@ -6,8 +6,8 @@ Copyright (c) 2022 Aiven, Helsinki, Finland. https://aiven.io/
 See LICENSE for details
 """
 
+from ..common.models import StorageModel
 from ..errors import FileNotFoundFromStorageError, InvalidConfigurationError, StorageError
-from ..models import StorageModel
 from ..notifier.interface import Notifier
 from .base import (
     BaseTransfer,
